% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextAction.R
\name{nextAction}
\alias{nextAction}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
nextAction(end, df_main, grabEnd, searchID, sub, phase)
}
\description{
Does not explicitely return data but assigns it to global variable inside the function.
}
