% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuatDistance.R
\name{QuatDistance}
\alias{QuatDistance}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
QuatDistance(frameGlobal)
}
\arguments{
\item{frameGlobal}{frame no for which distance should be computed (in relation to frame before)}
}
\description{
Read in data files for Object Sorting Task in VR.
}
