% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
readData(path, main = T, memory = T)
}
\arguments{
\item{path}{directory in which data is stored.}

\item{main}{indicates if main files (distractor task) should be loaded}

\item{memory}{indicates if memory files should be loaded

Does not explicitely return data but assigns it to global variable inside the function.}
}
\description{
Read in data files for Object Sorting Task in VR.
}
