% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFrames.R
\name{searchFrames}
\alias{searchFrames}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
searchFrames(df_main, cutoff = 10)
}
\description{
Does not explicitely return data but assigns it to global variable inside the function.
}
