% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceFramesWM.R
\name{sequenceFramesWM}
\alias{sequenceFramesWM}
\title{Get start and end of WM sequences. Sequences always starts with foxation on model.
Sequences ends when new model fixationis detected.}
\usage{
sequenceFramesWM(df_main, cutoff = 10)
}
\arguments{
\item{df_main}{combined data frame of main VR task}

\item{cutoff}{How many frames need to be between two fixations on model to count as two fixations?}
}
\value{
nested list. Each display is stored as a seperate list with global frame/row number of sequence start. Last value
is end frame of dispplay for easier looping later.
}
\description{
--> function makes sure to divide into subs/timepoint/session/phase/displays
to avoid spill over effects
}
\details{
returns nested list (for each display to avoid spill over!)
last frame number in each list is last frame + 1 (makes looping later easier)
}
