% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingPlacement.R
\name{SequencingPlacement}
\alias{SequencingPlacement}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
SequencingPlacement(df)
}
\description{
Does not explicitely return data but assigns it to global variable inside the function.
}
