% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingSearch.R
\name{SequencingSearch}
\alias{SequencingSearch}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
SequencingSearch(df_main, searchFrames, WM_dat, cutoff = 2)
}
\description{
Does not explicitely return data but assigns it to global variable inside the function.
}
