% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencingWM.R
\name{SequencingWM}
\alias{SequencingWM}
\title{Read in data files for Object Sorting Task in VR.}
\usage{
SequencingWM(df_main, sequenceFrames, cutoff = 2)
}
\description{
Does not explicitely return data but assigns it to global variable inside the function.
}
